refactor: moving to __init__
