+++ src/image_go_nord_client/main.py
@@ -1,41 +1,18 @@
 #! /usr/bin/env python3
 
 import logging
-import argparse
 from pathlib import Path
 import sys
 
 from os import listdir
 from typing import Union
 from ImageGoNord import GoNord
-from argparse import RawDescriptionHelpFormatter
 
-__doc__ = """ImageGoNord, a converter for a rgb images to norththeme palette.
-Usage: gonord [OPTION]...
-
-Mandatory arguments to long options are mandatory for short options too.
-
-Startup:
-
-
-Theme options:
-  --PALETTE[=LIST_COLOR_SET]        the palettes can be found in the
-                                    src/palettes/ directory (actually there is
-                                    only nord), by replacing 'PALETTE' with a
-                                    theme name it is possible to select it. If
-                                    necessary you can specify the set of colors
-                                    you want to use.
-                                    Ex: python src/cli.py --nord=Aurora,PolarNight,SnowStorm
-                                    Ex: python src/cli.py --monokai
-
-Conversion:
-
-  -pa=INT,INT, --pixel-area=INT,INT specify pixels of the area for average
-                                    color calculation
-
-Email bug reports, questions, discussions to <schrodinger.hat.show@gmail.com>
-and/or open issues at https://github.com/Schrodinger-Hat/ImageGoNord/issues/new
-"""
+from image_go_nord_client import (
+    get_argument_parser,
+    get_default_palette_path,
+    get_palette_list,
+)
 
 
 class confarg:
@@ -52,10 +29,6 @@ class confarg:
     }
 
 
-VERSION = (Path(__file__).parent / "VERSION").read_text().strip()
-DEFAULT_EXTENSION = ".png"
-OUTPUT_IMAGE_NAME = "nord" + DEFAULT_EXTENSION
-
 __ALL__ = ["to_console", "get_version", "main"]
 
 
@@ -69,102 +42,14 @@ def to_console(quiet_mode, *params):
 
 logging.basicConfig(level=logging.INFO, format="[%(levelname)s] %(message)s")
 
-parser = argparse.ArgumentParser(
-    description=__doc__,
-    add_help=True,
-    prog="image-go-nord-client",
-    formatter_class=RawDescriptionHelpFormatter,
-)
-
-parser.add_argument(
-    "-v",
-    "--version",
-    action="version",
-    version=VERSION,
-    help="show version number and exit",
-)
-
-parser.add_argument(
-    "-i",
-    "--img",
-    type=str,
-    dest="input_path",
-    metavar="PATH",
-    required=True,
-    help="specify input image path",
-)
-
-parser.add_argument(
-    "-o",
-    "--out",
-    type=str,
-    dest="output_path",
-    metavar="PATH",
-    default=OUTPUT_IMAGE_NAME,
-    help="specify output image path",
-)
-
-
-parser.add_argument(
-    "-q",
-    "--quiet",
-    action="store_true",
-    dest="quiet_mode",
-    default=False,
-    help="quiet (no output)",
-)
-
-parser.add_argument(
-    "-na",
-    "--no-avg",
-    action="store_true",
-    dest="disable_avg_pixels",
-    default=False,
-    help="do not use the average pixels optimization algorithm on conversion",
-)
-
-parser.add_argument(
-    "-b",
-    "--blur",
-    action="store_true",
-    dest="enable_blur",
-    default=False,
-    help="use blur on the final result",
-)
-
-def parse_pixels_area(value: str):
-    if not value:
-        raise TypeError( "Invalid value for pixels area: {}".format(value))
-    
-    values = value.split(",") 
-    if len(values) > 2:
-        raise ValueError( "Invalid number of parameters for pixels area: {}".format(value))
-    
-    if not all(map(lambda x: x.isdigit(), values)):
-        raise ValueError( "Invalid value for pixels area, all should be integer: {}".format(value))
-    
-    return values
-
-parser.add_argument(
-    "-pa",
-    "--pixels-area",
-    type=parse_pixels_area,
-    dest="pixels_area",
-    metavar="INT[,INT]",
-    default=[],
-    help="specify pixels of the area for average color calculation",
-)
-
 
 def main(argv: Union[list[str], None] = None):
-    global OUTPUT_IMAGE_NAME
-
-    output_image_path = OUTPUT_IMAGE_NAME
     PALETTE_CHANGED = False
 
     if argv is None:
         argv = sys.argv.copy()
 
+    parser = get_argument_parser()
     arguments, uknown_args = parser.parse_known_args(argv.copy())
     if arguments.quiet_mode:
         logging.basicConfig(level=logging.CRITICAL)
@@ -192,12 +77,10 @@ def main(argv: Union[list[str], None] = None):
         logging.info("Set up pixels width area: %s", w)
         logging.info("Set up pixels height area: %s", h)
 
-    # Get absolute path of source project
     src_path = Path(__file__).parent
+    palettes = get_palette_list()
 
-    # Get all palettes created
-    palettes = [folder.name.lower() for folder in (src_path / "palettes").iterdir()]
-
+    PALETTE_CHANGED = False
     for arg in uknown_args:
         key_value = [kv for kv in arg.split("=", 1) if kv != ""]
         key = key_value[0].lower()
@@ -263,7 +146,7 @@ def main(argv: Union[list[str], None] = None):
 
     if not PALETTE_CHANGED:
         to_console(arguments.quiet_mode, confarg.logs["pals"][4])
-        palette_path = src_path / "palettes" / "Nord"
+        palette_path = Path(get_default_palette_path())
         go_nord.reset_palette()
         palette_set = [
             palette_file.name.replace(".txt", "")
